refact: move function for keys manipulation outside of api package
These helper functions are used by multiple modules and there should not
be any circular dependencies
